WU-21 Create Homeuser.java
The Homeuser includes a listener for a spinner (userTypeSpinnerad) to detect the selected item. If the selected item is "ASE subject," it retrieves data of the content related to ASE from Firebase Realtime Database based on certain conditions and updates a RecyclerView (postRecyclerView) with the retrieved data for the "ASE" subject.